backend/requirements.txt
asgiref==3.7.2
Django==5.0
django-cors-headers==4.3.1
django-js-asset==2.2.0
django-rest-framework==0.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
Pillow==10.1.0
PyJWT==2.8.0
pytz==2023.3.post1
sqlparse==0.4.4
typing_extensions==4.9.0
